Add tests for WithIsFocused
